+++ docs/changes.rst
@@ -5,6 +5,7 @@ Weblate 5.16.1
 
 .. rubric:: New features
 
+* Added :setting:`WEBSITE_ALERTS_ENABLED` setting to allow disabling website availability checks and alerts.
 * :doc:`/formats/asciidoc` is now a supported file format.
 * Added :setting:`REGISTRATION_ALLOW_DISPOSABLE_EMAILS` to optionally allow disposable e-mail domains during registration (Docker env: :envvar:`WEBLATE_REGISTRATION_ALLOW_DISPOSABLE_EMAILS`).